+++ Base32768.Test/Base32768Test.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using FluentAssertions;
+using Kzrnm.Convert.Base32768.Models;
 using Xunit;
 
 namespace Kzrnm.Convert.Base32768
