+++ Base32768.Test/Models/IsExternalInit.cs
